SCIA/requirements.txt
langchain==0.1.19
langchain-experimental==0.0.58
langchain-openai==0.1.6
numpy==1.26.4
openai==1.28.0
openai-whisper==20231117
pandas==2.2.2
pynput==1.7.7
python-dotenv==1.0.1
sounddevice==0.4.6
soundfile==0.12.1
tabulate==0.9.0
